"Fools" Interactive Music Video Installation
An interactive music video installation made together with Sultanali
Barodawala including a performance of "Fools We Mortals Be" by Etta
James, using the Kinect to get the dancers in the video to dance with
the audience. The two prototypes below explore new avenues for
musicians to advertise themselves and transfer the experience of a live
music performance. Music by Sean Steele, Blair Milne on vibes, Joe
Ziegler on drums, and Jayne Fife on bass. On point we had Emily Katz
and Katy Roth.